Now using only token to get all user info, Getting an Exemplaire from server to add to cart, Updating cart to get only one item with same ID, Adding last component 'location'

+++ src/app/app.component.ts
@@ -55,7 +55,13 @@ export class AppComponent implements OnInit {
   }
 
   onClick(idExemplaire): void {
-    this.cartService.removeFromCart(idExemplaire);
-    window.alert('item retiré');
+    if (window.confirm('Etes vous sûr de supprimer cet item?')){
+      this.cartService.removeFromCart(idExemplaire);
+      window.alert('item retiré');
+    }
+  }
+
+  onReserver(): void {
+    this.router.navigate(['/reservation']);
   }
 }
